fear: rating refactored
